refactor(webview): split desktop.rs into sub-modules (Iteration #143)
- Extract icon.rs: load_window_icon function
- Extract event_loop.rs: run_desktop function
- Extract window_builder.rs: create_window_and_event_loop helper
- Extract webview_builder.rs: configure_webview_builder helper
- desktop/mod.rs now 879 lines (under 1000 limit)
- All tests pass, clippy clean
